Reduce the amount of query logs. Queries are large and there's no point to log it more than once
