+++ packages/hierarchy-builder/src/hierarchy-builder/internal/TreeNodesReader.ts
@@ -5,13 +5,11 @@
 
 import { INodeParser } from "../HierarchyDefinition";
 import { InstanceHierarchyNodeProcessingParams, ParsedHierarchyNode, ParsedInstanceHierarchyNode } from "../HierarchyNode";
-import { getLogger } from "../Logging";
 import { ECSqlQueryDef } from "../queries/ECSqlCore";
 import { ILimitingECSqlQueryExecutor } from "../queries/LimitingECSqlQueryExecutor";
 import { NodeSelectClauseColumnNames } from "../queries/NodeSelectQueryFactory";
 import { ConcatenatedValue } from "../values/ConcatenatedValue";
 import { Id64String } from "../values/Values";
-import { LOGGING_NAMESPACE } from "./Common";
 
 /** @internal */
 export interface TreeQueryResultsReaderProps {
@@ -30,7 +28,6 @@ export class TreeQueryResultsReader {
   }
 
   public async *read(queryExecutor: ILimitingECSqlQueryExecutor, query: ECSqlQueryDef, limit?: number | "unbounded"): AsyncGenerator<ParsedHierarchyNode> {
-    getLogger().logInfo(`${LOGGING_NAMESPACE}.TreeQueryResultsReader`, `Executing query: ${query.ecsql}`);
     for await (const row of queryExecutor.createQueryReader(query, { rowFormat: "ECSqlPropertyNames", ...(limit !== undefined ? { limit } : undefined) })) {
       yield this._props.parser(row);
     }
