+++ packages/hierarchy-builder/src/presentation-hierarchy-builder.ts
@@ -34,7 +34,7 @@ export * from "./hierarchy-builder/HierarchyNode";
 export * from "./hierarchy-builder/HierarchyProvider";
 export * from "./hierarchy-builder/Logging";
 export * from "./hierarchy-builder/Metadata";
-export * from "./hierarchy-builder/Utils";
+export { OmitOverUnion } from "./hierarchy-builder/Utils";
 export * from "./hierarchy-builder/queries/ECSqlCore";
 export * from "./hierarchy-builder/queries/InstanceLabelSelectClauseFactory";
 export { createLimitingECSqlQueryExecutor, ILimitingECSqlQueryExecutor } from "./hierarchy-builder/queries/LimitingECSqlQueryExecutor";
