--- packages/hierarchy-builder/src/test/queries/Utils.ts
@@ -1,18 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
-
-export function trimWhitespace(str: string): string;
-export function trimWhitespace(str: string | undefined): string | undefined;
-export function trimWhitespace(str: string | undefined): string | undefined {
-  if (!str) {
-    return str;
-  }
-  return str
-    .replaceAll(/\s+/gm, " ") // replace all consecutive spaces with a single space
-    .replaceAll(/\(\s+/g, "(") // remove spaces after opening parentheses
-    .replaceAll(/\s+\)/g, ")") // remove spaces before closing parentheses
-    .replaceAll(/\s+,/g, ",") // remove spaces before comma
-    .trim(); // remove spaces from beginning and end of the string
-}
